Add FileBufferedReaderService class and unit test
The FileBufferedReaderService class simple returns a BufferedReader
using the ClassPathResource bean, and the file specificed in the
application.yml through the readFile() method. This isolates the IO
logic outside of the handler class, and makes it easy to unit test. Unit
test is also included and simply returns a BufferedReader from a simple
"foobarinputstream" string.