Add SourceApplicationRunner class and unit/int test
The SourceApplicationRunner uses the Source.class and
@InboundChannelAdapter to identify it as a source and also push data in
JSON string format to the Source.OUTPUT channel. It uses the
FileBufferedReaderServer to get a BufferedReader from the
starbucks_us_locations.csv file where we can read the file line by line.

We JSONify each line and then pass them one by one to the source output
channel. The integration test uses Spring Cloud Stream Test Support
suite in order to test that we are getting at least the first two lines
in the output channel. We also add a unit test.

Note that we return null at the end of the method, where the handle
signature returns a string. We cannot use the @StreamListener annotation
with a source application, so we must use the @InboundChannelAdapter
annotation which requires us to return a value. We return null instead,
and push our JSONified string data to the output channel ourselves.