Add test ActiveProfiles to integration tests
We add a 'test' profile to each of our integration tests so that we can
use our embedded mongo service. Since we have to override a bean in
order to connect to a mongo repository with version greater than 3.x, we
don't want to instantiate this within the test. We don't want it
instantiated in the test because it will try to reach out to a docker
version of mongo, not the embedded version.